requirements.txt
streamlit==1.27.2
sentence-transformers==2.2.2
pypdf2==3.0.1
pinecone-client==2.2.2
cohere==4.6.0
numpy==1.23.5
scikit-learn==1.2.2
notebook==7.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0
